Process information

Get or set attributes of the current process, with the same names, arguments and results as the 'unix' package.

Usage

getuid()

getgid()

geteuid()

getegid()

getpid()

getppid()

getpgid()

getpriority()

setuid(uid)

seteuid(uid)

setgid(gid)

setegid(gid)

setpgid(pgid = 0)

setpriority(prio)

kill(pid, signal = SIGTERM)

Arguments

uid

User id.

gid

Group id.

pgid

Process group id; 0 uses the current process id.

prio

Priority.

pid

Process id.

signal

Signal number, tools::SIGTERM by default.

Value

The getters and setters return the (new) value as an integer; kill() returns NULL.

Details

pid: process id; ppid: parent process id; pgid: process group id.

uid, euid: real and effective user id; gid, egid: group ids.

prio: scheduling priority; a higher value is a lower priority.

An unprivileged process cannot change its uid and can only lower its priority (raise the value).

References

GETUID(2) GETPID(2) GETPGID(2) GETPRIORITY(2)

Examples

getuid()
#> [1] 1001
getpid()
#> [1] 6462
getpriority()
#> [1] 0

# Changing the process is irreversible, so shown in throwaway children:
eval_fork(setpriority(getpriority() + 1))
#> [1] 1
eval_fork(setpgid())
#> [1] 6804
eval_fork(getppid()) == getpid()
#> [1] TRUE